introduce a separate file for AnnotatorHost
The various JSDoc-annotation modules all depend intimately and
circularly on each other and the AnnotatorHost.  There's not a
particularly good place for this to live in another module, so
move it to its own file to break the circularity.
